Running version of SA that works
This includes 4 versions of the SA (sd. based on lit or %of mean x either running SA on all params at the same time or one by one)

Need to verify: The values used for stdev from literature.

Also possible next version: change the cost_pup systematically in the functions, but this is not essential.