Updating Variation CSV
I am trying to condense the variations from literature into values for each parameter (by sex, age, etc.). I am having trouble getting the activity budget results to add up to 100 (they are % of the day spent in each behaviour).

I think I should be able to fix it in the csv file by altering the proportions, but I cant figure out what I am doing wrong, because even the easy ones that are only an average of a couple of them dont add up- like the juvenile activity budgets.

or I need to stop trying to average them and I need just add the activity budget values, because some of them do just need to be added together and not averaged- for example- swimming, grooming, and other all make up (non-foraging) "activity", so they need to be combined to equal activity, but proportionally, but the code keeps just averaging those three values, and i cant figure out how to change the code.